implement configurable supervision tree
Currently pigeon causes the entire BEAM to restart when an APNS
key is revoked or a cert expires.
This change allows the library user to be more in control of what
happens when pigeon crashes by allowing the library user to
insert the pigeon supervisor in the library user's supervision tree.
This change still only allows one pigeon supervisor in an application.
A name conflict will occur if trying to add pigeon more than once.